[test] move mining helper functions into test library
