+++ src/bench/block_assemble.cpp
@@ -6,6 +6,7 @@
 #include <consensus/validation.h>
 #include <crypto/sha256.h>
 #include <test/util.h>
+#include <test/util/mining.h>
 #include <txmempool.h>
 #include <validation.h>
