+++ src/bench/wallet_balance.cpp
@@ -7,6 +7,7 @@
 #include <node/context.h>
 #include <optional.h>
 #include <test/util.h>
+#include <test/util/mining.h>
 #include <validationinterface.h>
 #include <wallet/wallet.h>
